backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pytest>=9.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
python-jose[cryptography]>=3.3.0
